backend/requirements.txt
# ── Core ML (Milestone 1) ──────────────────────────────────────────────────
numpy
pandas
scikit-learn
nltk
joblib
matplotlib
seaborn

# ── API ────────────────────────────────────────────────────────────────────
fastapi
uvicorn
pydantic
requests
python-dotenv

# ── Scraping ───────────────────────────────────────────────────────────────
newspaper3k
lxml_html_clean
beautifulsoup4
requests

# ── RAG / Vector DB (Milestone 2) ─────────────────────────────────────────
faiss-cpu
sentence-transformers

# ── LLM (Milestone 2) ─────────────────────────────────────────────────────
# HuggingFace Inference API — no local GPU needed
huggingface-hub

# ── PDF Export (Milestone 2) ───────────────────────────────────────────────
reportlab

# ── Notebooks ─────────────────────────────────────────────────────────────
jupyter
ipykernel
